fix(assisted): adjust interfaces names and remove unnecessary code
